requirements.txt
discord.py
Flask
python-dotenv
aiofiles
motor
pymongo
waitress
